last separations
