+++ lib/src/metta/runner/stdlib/debug.rs
@@ -1,11 +1,22 @@
 use crate::*;
 use crate::metta::*;
 use crate::metta::text::Tokenizer;
+use crate::space::*;
+use crate::common::assert::vec_eq_no_order;
 
-use crate::metta::runner::stdlib::{grounded_op, atom_to_string, regex};
+use crate::metta::runner::stdlib::{grounded_op, atom_to_string, regex, interpret_no_error, unit_result};
 
 use std::convert::TryInto;
 
+
+fn assert_results_equal(actual: &Vec<Atom>, expected: &Vec<Atom>, atom: &Atom) -> Result<Vec<Atom>, ExecError> {
+    log::debug!("assert_results_equal: actual: {:?}, expected: {:?}, actual atom: {:?}", actual, expected, atom);
+    let report = format!("\nExpected: {:?}\nGot: {:?}", expected, actual);
+    match vec_eq_no_order(actual.iter(), expected.iter()) {
+        Ok(()) => unit_result(),
+        Err(diff) => Err(ExecError::Runtime(format!("{}\n{}", report, diff)))
+    }
+}
 
 /// Implement trace! built-in.
 ///
@@ -93,16 +104,186 @@ impl CustomExecute for PrintAlternativesOp {
     }
 }
 
-pub fn register_runner_tokens(tref: &mut Tokenizer) {
+#[derive(Clone, Debug)]
+pub struct AssertEqualOp {
+    space: DynSpace,
+}
+
+grounded_op!(AssertEqualOp, "assertEqual");
+
+impl AssertEqualOp {
+    pub fn new(space: DynSpace) -> Self {
+        Self{ space }
+    }
+}
+
+impl Grounded for AssertEqualOp {
+    fn type_(&self) -> Atom {
+        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM])
+    }
+
+    fn as_execute(&self) -> Option<&dyn CustomExecute> {
+        Some(self)
+    }
+}
+
+impl CustomExecute for AssertEqualOp {
+    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
+        log::debug!("AssertEqualOp::execute: {:?}", args);
+        let arg_error = || ExecError::from("assertEqual expects two atoms as arguments: actual and expected");
+        let actual_atom = args.get(0).ok_or_else(arg_error)?;
+        let expected_atom = args.get(1).ok_or_else(arg_error)?;
+
+        let actual = interpret_no_error(self.space.clone(), actual_atom)?;
+        let expected = interpret_no_error(self.space.clone(), expected_atom)?;
+
+        assert_results_equal(&actual, &expected, actual_atom)
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct AssertEqualToResultOp {
+    space: DynSpace,
+}
+
+grounded_op!(AssertEqualToResultOp, "assertEqualToResult");
+
+impl AssertEqualToResultOp {
+    pub fn new(space: DynSpace) -> Self {
+        Self{ space }
+    }
+}
+
+impl Grounded for AssertEqualToResultOp {
+    fn type_(&self) -> Atom {
+        Atom::expr([ARROW_SYMBOL, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM, ATOM_TYPE_ATOM])
+    }
+
+    fn as_execute(&self) -> Option<&dyn CustomExecute> {
+        Some(self)
+    }
+}
+
+impl CustomExecute for AssertEqualToResultOp {
+    fn execute(&self, args: &[Atom]) -> Result<Vec<Atom>, ExecError> {
+        log::debug!("AssertEqualToResultOp::execute: {:?}", args);
+        let arg_error = || ExecError::from("assertEqualToResult expects two atoms as arguments: actual and expected");
+        let actual_atom = args.get(0).ok_or_else(arg_error)?;
+        let expected = TryInto::<&ExpressionAtom>::try_into(args.get(1).ok_or_else(arg_error)?)
+            .map_err(|_| arg_error())?
+            .children();
+
+        let actual = interpret_no_error(self.space.clone(), actual_atom)?;
+
+        assert_results_equal(&actual, &expected.into(), actual_atom)
+    }
+}
+
+pub fn register_runner_tokens(tref: &mut Tokenizer, space: &DynSpace) {
     let trace_op = Atom::gnd(TraceOp{});
     tref.register_token(regex(r"trace!"), move |_| { trace_op.clone() });
     let print_alternatives_op = Atom::gnd(PrintAlternativesOp{});
     tref.register_token(regex(r"print-alternatives!"), move |_| { print_alternatives_op.clone() });
+    let assert_equal_op = Atom::gnd(AssertEqualOp::new(space.clone()));
+    tref.register_token(regex(r"assertEqual"), move |_| { assert_equal_op.clone() });
+    let assert_equal_to_result_op = Atom::gnd(AssertEqualToResultOp::new(space.clone()));
+    tref.register_token(regex(r"assertEqualToResult"), move |_| { assert_equal_to_result_op.clone() });
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::metta::runner::{Metta, EnvBuilder, SExprParser};
+    use crate::common::test_utils::metta_space;
+
+    use regex::Regex;
+
+    fn assert_runtime_error(actual: Result<Vec<Atom>, ExecError>, expected: Regex) {
+        match actual {
+            Err(ExecError::Runtime(msg)) => assert!(expected.is_match(msg.as_str()),
+                                                    "Incorrect error message:\nexpected: {:?}\n  actual: {:?}", expected.to_string(), msg),
+            _ => assert!(false, "Error is expected as result, {:?} returned", actual),
+        }
+    }
+
+    #[test]
+    fn metta_assert_equal_op() {
+        let metta = Metta::new(Some(EnvBuilder::test_env()));
+        let assert = AssertEqualOp::new(metta.space().clone());
+        let program = "
+            (= (foo $x) $x)
+            (= (bar $x) $x)
+        ";
+        assert_eq!(metta.run(SExprParser::new(program)), Ok(vec![]));
+        assert_eq!(metta.run(SExprParser::new("!(assertEqual (foo A) (bar A))")), Ok(vec![
+            vec![UNIT_ATOM],
+        ]));
+        assert_eq!(metta.run(SExprParser::new("!(assertEqual (foo A) (bar B))")), Ok(vec![
+            vec![expr!("Error" ({assert.clone()} ("foo" "A") ("bar" "B")) "\nExpected: [B]\nGot: [A]\nMissed result: B")],
+        ]));
+        assert_eq!(metta.run(SExprParser::new("!(assertEqual (foo A) Empty)")), Ok(vec![
+            vec![expr!("Error" ({assert.clone()} ("foo" "A") "Empty") "\nExpected: []\nGot: [A]\nExcessive result: A")]
+        ]));
+    }
+
+    #[test]
+    fn metta_assert_equal_to_result_op() {
+        let metta = Metta::new(Some(EnvBuilder::test_env()));
+        let assert = AssertEqualToResultOp::new(metta.space().clone());
+        let program = "
+            (= (foo) A)
+            (= (foo) B)
+            (= (bar) C)
+            (= (baz) D)
+            (= (baz) D)
+        ";
+        assert_eq!(metta.run(SExprParser::new(program)), Ok(vec![]));
+        assert_eq!(metta.run(SExprParser::new("!(assertEqualToResult (foo) (A B))")), Ok(vec![
+            vec![UNIT_ATOM],
+        ]));
+        assert_eq!(metta.run(SExprParser::new("!(assertEqualToResult (bar) (A))")), Ok(vec![
+            vec![expr!("Error" ({assert.clone()} ("bar") ("A")) "\nExpected: [A]\nGot: [C]\nMissed result: A")],
+        ]));
+        assert_eq!(metta.run(SExprParser::new("!(assertEqualToResult (baz) (D))")), Ok(vec![
+            vec![expr!("Error" ({assert.clone()} ("baz") ("D")) "\nExpected: [D]\nGot: [D, D]\nExcessive result: D")]
+        ]));
+    }
+
+    #[test]
+    fn assert_equal_op() {
+        let space = DynSpace::new(metta_space("
+            (= (foo) (A B))
+            (= (foo) (B C))
+            (= (bar) (B C))
+            (= (bar) (A B))
+            (= (err) (A B))
+        "));
+
+        let assert_equal_op = AssertEqualOp::new(space);
+
+        assert_eq!(assert_equal_op.execute(&mut vec![expr!(("foo")), expr!(("bar"))]), unit_result());
+
+        let actual = assert_equal_op.execute(&mut vec![expr!(("foo")), expr!(("err"))]);
+        let expected = Regex::new("\nExpected: \\[(A B)\\]\nGot: \\[\\((B C)|, |(A B)\\){3}\\]\nExcessive result: (B C)").unwrap();
+        assert_runtime_error(actual, expected);
+
+        let actual = assert_equal_op.execute(&mut vec![expr!(("err")), expr!(("foo"))]);
+        let expected = Regex::new("\nExpected: \\[\\((B C)|, |(A B)\\){3}\\]\nGot: \\[(A B)\\]\nMissed result: (B C)").unwrap();
+        assert_runtime_error(actual, expected);
+    }
+
+    #[test]
+    fn assert_equal_to_result_op() {
+        let space = DynSpace::new(metta_space("
+            (= (foo) (A B))
+            (= (foo) (B C))
+        "));
+        let assert_equal_to_result_op = AssertEqualToResultOp::new(space);
+
+        assert_eq!(assert_equal_to_result_op.execute(&mut vec![
+            expr!(("foo")), expr!(("B" "C") ("A" "B"))]),
+                   unit_result());
+    }
 
     #[test]
     fn trace_op() {
